refactor: :art: componetizado y creacion del adapter para transformar la data de UserReviewComment, aun necesita los types y en reviews hay que separar el context y el custom hook en su respectiva carpeta
