+++ src/components/details/reviews/CommentIterator.tsx
@@ -1,11 +1,6 @@
-import { formatDate } from "../../../helper";
 import { useState } from "react";
-import {
-  ButtonShowAllReviews,
-  ReviewsDetails,
-  StarReviewRating,
-} from "../../../components";
-import type { IReview } from "../../../interfaces";
+import { ButtonShowAllReviews, UserReviewComment } from "@/components";
+import type { IReview } from "@/interfaces";
 
 export const CommentIterator = ({ reviews }: { reviews: IReview[] }) => {
   const [ShowAllReviews, setShowAllReviews] = useState(false);
@@ -13,21 +8,7 @@ export const CommentIterator = ({ reviews }: { reviews: IReview[] }) => {
   return (
     <>
       {reviews.slice(0, 2).map((review) => (
-        <ReviewsDetails key={review.user.id}>
-          <ReviewsDetails.Container>
-            <ReviewsDetails.UserAvatar imageUrl={review.user.imageUrl ?? ""}>
-              <ReviewsDetails.Username name={review.user.name}>
-                <ReviewsDetails.UserRole title={review.user.title ?? ""} />
-              </ReviewsDetails.Username>
-            </ReviewsDetails.UserAvatar>
-
-            <ReviewsDetails.PostedTime date={formatDate(review.createAt)}>
-              <StarReviewRating rating={review.rating} />
-            </ReviewsDetails.PostedTime>
-          </ReviewsDetails.Container>
-
-          <ReviewsDetails.UserComment text={review.text} />
-        </ReviewsDetails>
+        <UserReviewComment key={review.user.id} review={review} />
       ))}
       {totalReviews > 0 && (
         <ButtonShowAllReviews
@@ -36,23 +17,9 @@ export const CommentIterator = ({ reviews }: { reviews: IReview[] }) => {
         />
       )}
       {ShowAllReviews &&
-        reviews.slice(2).map((review) => (
-          <ReviewsDetails key={review.user.id}>
-            <ReviewsDetails.Container>
-              <ReviewsDetails.UserAvatar imageUrl={review.user.imageUrl ?? ""}>
-                <ReviewsDetails.Username name={review.user.name}>
-                  <ReviewsDetails.UserRole title={review.user.title ?? ""} />
-                </ReviewsDetails.Username>
-              </ReviewsDetails.UserAvatar>
-
-              <ReviewsDetails.PostedTime date={formatDate(review.createAt)}>
-                <StarReviewRating rating={review.rating} />
-              </ReviewsDetails.PostedTime>
-            </ReviewsDetails.Container>
-
-            <ReviewsDetails.UserComment text={review.text} />
-          </ReviewsDetails>
-        ))}
+        reviews
+          .slice(2)
+          .map((review) => <UserReviewComment key={review.user.id} review={review} />)}
     </>
   );
 };
